3496 - Rename GeoMap to Map
