--- src/client/pages/Geo/GeoMap/index.ts
@@ -1,1 +1,0 @@
-export { default } from './GeoMap'
