+++ src/client/pages/Geo/Map/hooks/useCountSectionSelectedLayers.ts
